[FEAT] record 조회 구현, 업데이트 구현중..

+++ src/main/java/com/d204/rumeet/record/model/mapper/RecordMapper.java
@@ -6,7 +6,7 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface RecordMapper {
 
-    void updateRecord(int userId, int pace, int km);
+    void updateRecord(int userId, float pace, float km);
 
     RecordDto getRecord(int userId);
 }
